Logging try-catch

+++ controllers/api.js
@@ -8,6 +8,7 @@ const logger = require("../functions/logger");
 var Client = require("../models/client");
 const bcrypt = require("bcrypt");
 const isAuth = require("../middleware/isAuth");
+const isAdmin = require("../middleware/isAdmin");
 const isNotAuth = require("../middleware/isNotAuth");
 const short = require("short-uuid");
 const API = require("./payment.js");
@@ -63,26 +64,30 @@ router.post("/login", isNotAuth, async function (req, res) {
   }
 });
 
-router.get("/deletetest", isAuth, async function (req, res) {
-  await Client.deleteMany({});
-  // await Client.deleteOne({ name: "Jane Doe" });
-  // await Client.deleteOne({ name: "Sourav" });
-  await User.deleteMany({});
-  const testUser = new User();
-  testUser.username = "shahina";
-  testUser.password = await bcrypt.hash("testpassword", 10);
-  await testUser.save();
-  const x = new User();
-  x.username = "sagar";
-  x.password = await bcrypt.hash("password", 10);
-  await x.save();
-  await Logs.deleteMany({});
-  // await User.findOneAndUpdate({ username: "sagar" }, { $set: { clients: [] } });
-  // let results = await User.find();
-  // let results = await Token.find();
-  // let results = await Token.deleteMany({});
-  let results = await Client.find();
-  res.send(results);
+router.get("/cleandb", isAdmin, async function (req, res) {
+  if (req.session.username == "sagar") {
+    await Client.deleteMany({});
+    // await Client.deleteOne({ name: "Jane Doe" });
+    // await Client.deleteOne({ name: "Sourav" });
+    await User.deleteMany({});
+    const testUser = new User();
+    testUser.username = "shahina";
+    testUser.password = await bcrypt.hash("testpassword", 10);
+    await testUser.save();
+    const x = new User();
+    x.username = "sagar";
+    x.password = await bcrypt.hash("password", 10);
+    await x.save();
+    await Logs.deleteMany({});
+    // await User.findOneAndUpdate({ username: "sagar" }, { $set: { clients: [] } });
+    // let results = await User.find();
+    // let results = await Token.find();
+    // let results = await Token.deleteMany({});
+    let results = await Client.find();
+    res.send(results);
+  } else {
+    res.send("Unauthorized");
+  }
 });
 
 router.get("/test", async function (req, res) {
